Simplify nameOverride construction
Prior to this change, nameOverride construction for proliferated
contexts was opinionated and over-complicated.  With the introduction of
the `proliferationStack` value in the previous commit, this complexity
could easily be recreated by a parent chart, leaving the default
behaviour to be more simple.

This change moves to using a simple suffixing approach to `nameOverride`
construction, expecting parent charts to update their name-related
partial templates if other patterns are required.