Include explicit nameOverrides in proliferationStack
Prior to this change, more complex name construction was possible, but
in so doing would lose information about explicit nameOverrides at
different levels of a proliferation stack.

This change includes any nameOverride explicitly specified at a given
level in the proliferationStack, thereby facilitating their use in the
recreation of the default behaviour, which is a minumum functionality
for name construction.